+++ src/routes/demo/rbac-auth/onboarding.tsx
@@ -20,6 +20,7 @@ import {
 	CardTitle,
 } from "#/components/ui/card";
 import { useAppAuth } from "#/hooks/use-app-auth";
+import { hasPermission, isFairLendStaffAdmin } from "#/lib/auth";
 import {
 	ROLE_COLOR_CLASSES,
 	ROLE_DISPLAY_METADATA,
@@ -99,7 +100,9 @@ const STATE_MACHINE_NODES = [
 
 function OnboardingPage() {
 	const auth = useAppAuth();
-	const canReview = auth.permissions.includes("onboarding:review");
+	const canReview =
+		isFairLendStaffAdmin(auth) &&
+		hasPermission(auth.permissions, "onboarding:review");
 	const isMember =
 		auth.role === "member" || (auth.roles.length === 0 && !auth.loading);
